+++ app/src/js/utils/table-config/granules.js
@@ -69,8 +69,8 @@ export const tableColumns = [
   {
     Header: 'Error Type',
     accessor: (row) => get(row, 'error.Error', nullValue),
-    Cell: ({ row: { original } }) => (
-      (isEmpty(get(original, ['error']))) ? nullValue : get(original, 'error.Error', nullValue)
+    Cell: ({ row: { original: { error } } }) => (
+      (isEmpty(error) || /^("?)(None)\1$/.test(error.Cause)) ? nullValue : error.Error
     ),
     id: 'errorType',
     width: 100
@@ -80,7 +80,7 @@ export const tableColumns = [
     accessor: (row) => get(row, 'error.Cause', nullValue),
     id: 'error',
     Cell: ({ row: { original } }) => ( // eslint-disable-line react/prop-types
-      (isEmpty(get(original, ['error']))) ? nullValue : <ErrorReport report={get(original, 'error', nullValue)} truncate={true} disableScroll={true} />
+      (isEmpty(get(original, ['error'])) || /^("?)(None)\1$/.test(get(original, ['error', 'Cause']))) ? nullValue : <ErrorReport report={get(original, 'error', nullValue)} truncate={true} disableScroll={true} />
     ),
     disableSortBy: true,
     width: 175