+++ cypress/e2e/granules.cy.js
@@ -1005,6 +1005,39 @@ describe('Dashboard Granules Page', () => {
         .and('include', `collectionId=${encodeURIComponent(collectionId)}`)
         .and('include', `search=${searchShort}`);
     });
+
+    it('Should allow user to select execute workflow option from dropdown when from granule detail page', () => {
+      const granuleId = 'test_12345678_123456_metopa_12345_eps_o_coa_1234_ovwcl2';
+      const workflowNameTest = 'SecondTestWorkflow';
+      cy.intercept('GET', `/granules/${granuleId}*`).as('getGranule');
+      cy.intercept('PATCH', /\/granules\/.*/).as('executeWorkflow');
+
+      cy.visit('/granules');
+      cy.wait(1000);
+      cy.contains('.table .tbody .tr a', granuleId).then(($res) => {
+        expect($res).to.have.attr('href', `/granules/granule/${granuleId}`);
+        cy.wrap($res).click();
+      });
+      cy.wait('@getGranule');
+      cy.get('.heading--large').should('have.text', `Granule: ${granuleId}`);
+      cy.contains('button', 'Options').click();
+      cy.get('.dropdown__menu').contains('Execute').click();
+      cy.get('.modal-body .form__dropdown .dropdown__element input').as('workflow-input');
+      cy.get('@workflow-input').click({ force: true });
+      cy.get('.modal-body .form__dropdown .dropdown__element .react-select__menu').as('workflow-menu');
+      cy.get('@workflow-menu').should('be.visible');
+      cy.get('@workflow-menu').get('.react-select__option').each(($option) => {
+        if ($option.text() === workflowNameTest) {
+          cy.wrap($option).click();
+        }
+      });
+      cy.get('.modal-body .form__dropdown .dropdown__element .react-select__single-value').should('have.text', workflowNameTest);
+      cy.get('.button--submit').click();
+      cy.wait('@executeWorkflow').then(({ request }) => {
+        expect(request.body.action).to.equal('applyWorkflow');
+        expect(request.body.workflow).to.equal(workflowNameTest);
+      });
+    });
   });
 
   describe('when ESTIMATE_TABLE_ROW_COUNT is false', () => {
@@ -1157,7 +1190,7 @@ describe('Dashboard Granules Page', () => {
 
       cy.get('.button__apply-filter').click();
 
-      const noErrorGranuleID = 'test_12345678_123456_metopa_12345_eps_o_coa_1234_ovwcl2';
+      const noErrorGranuleID = 'MOD09GQ.A8022119.sk3Sph.006.0494433853533';
       const errorGranuleID = 'MOD09GQ.A2417309.YZ9tCV.006.4640974889044_ca2a8dfe';
       const unknownErrorGranuleID = 'coastal_12345678_123456_metopa_12345_eps_o_coa_1234_ovwcl3';
       cy.get(`[data-value="${noErrorGranuleID}"]`).children().as('noErrorColumns');
@@ -1167,12 +1200,12 @@ describe('Dashboard Granules Page', () => {
       cy.get('@errorColumns').eq(4).invoke('text').should('be.eq', 'UnexpectedFileSize');
       cy.get('@errorColumns').eq(5).invoke('text').should('match', /errorMessage/);
       cy.get(`[data-value="${unknownErrorGranuleID}"]`).children().as('unknownErrorColumns');
-      cy.get('@unknownErrorColumns').eq(4).invoke('text').should('be.eq', 'Unknown Error');
-      cy.get('@unknownErrorColumns').eq(5).invoke('text').should('match', /None/);
+      cy.get('@unknownErrorColumns').eq(4).invoke('text').should('be.eq', '--');
+      cy.get('@unknownErrorColumns').eq(5).invoke('text').should('be.eq', '--');
     });
 
     it('Should display error if available in granule overview page', () => {
-      const noErrorGranuleID = 'test_12345678_123456_metopa_12345_eps_o_coa_1234_ovwcl2';
+      const noErrorGranuleID = 'MOD09GQ.A8022119.sk3Sph.006.0494433853533';
       const errorGranuleID = 'MOD09GQ.A2417309.YZ9tCV.006.4640974889044_ca2a8dfe';
       const unknownErrorGranuleID = 'coastal_12345678_123456_metopa_12345_eps_o_coa_1234_ovwcl3';
       cy.visit(`/granules/granule/${errorGranuleID}`);
@@ -1184,8 +1217,7 @@ describe('Dashboard Granules Page', () => {
       cy.get('.error__report').should('not.exist');
       cy.visit(`/granules/granule/${unknownErrorGranuleID}`);
       cy.get('.heading--large').should('have.text', `Granule: ${unknownErrorGranuleID}`);
-      cy.get('.error__report').should('be.visible');
-      cy.get('.error__report').should('contain.text', 'None');
+      cy.get('.error__report').should('not.exist');
     });
   });
 });